Can now update the goalie database. Can work on a Report class now

+++ fantasy-hockey-hidden-gems-java/main/ca/ualberta/emar/Goalie.java
@@ -199,7 +199,7 @@ public class Goalie {
 		return totalGAA;
 	}
 	
-	protected void setTotalGAA(int totalGAA){
+	protected void setTotalGAA(float totalGAA){
 		this.totalGAA = totalGAA;
 	}
